# [WIP] Performance optimize Point protobuf message