Changes in Point message:

- Replace google.protobuf.Timestamp by int64 time in unix epoch nanoseconds.
- Replace oneof by a set of fields.

This is part 1 one of overall changes that yield about 2x performance improvement.

Benchmarking
============

Simple Go benchmark of *all* changes that I plan to propose demonstrates the following
improvement compared to OpenCensus Metrics encoding and decoding:

```
BenchmarkEncode/OpenCensus/Metrics-8    8	 660200437 ns/op	80432000 B/op	 1787000 allocs/op
BenchmarkEncode/OTLP/Metrics-8         19	 296717207 ns/op	49152000 B/op	    1000 allocs/op
BenchmarkDecode/OpenCensus/Metrics-8    5	1040309846 ns/op	458263993 B/op	11877000 allocs/op
BenchmarkDecode/OTLP/Metrics-8         12	 501670950 ns/op	346696049 B/op	 7124000 allocs/op
```

Encoding and decoding is about 2 times faster with new format. Memory usage is also
significantly smaller, number of allocations during decoding is roughly half of OpenCensus
which reflects the fact that many messages were eliminated and embedded inside others.

Benchmarks encode and decode 1000 batches of 2 metrics: in each batch 50 int64 Gauges with
5 time series and 50 Histograms of doubles with 1 time series and single bucket. Each time
series for both metrics contains 5 data points. Both metrics have 2 labels.

Benchmark source code is available at:
https://github.com/tigrannajaryan/exp-otelproto/blob/master/encodings/encoding_test.go