@tigrannajaryan 2019-10-25
Please do not merge this yet even if it gets approved. I am exploring an alternate representation for Point, which may be even better.