@jmacd 2019-10-25
I'm willing to go with what you feel is best, regarding the 0 vs boolean question.