@tigrannajaryan 2019-10-30
I am closing this one and will open a series of separate PRs with incremental improvements.